Invisibility Cloak using Python
This is a beginner friendly project for
people who want to start working in Computer Vision.
By using OpenCV library, we can create a real time,
invisibility cloak. I have used the values for
masking red color. So we will be using a red colored
cloth. It is a fun project and also great for diving into
concepts of Computer Vision and Image Processing